Refactor JsonModelsInterface
